Main.js and Login.js in progress. React-redux dependence was added
