+++ package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.1",
     "bcrypt": "^5.1.0",
+    "react-redux": "^8.0.5",
     "react-router": "^6.4.4"
   }
 }
